feat: move LoginErrorManager and TranslationConfig to commons (batch 20 part 4)
Additional moves:
- LoginErrorManager (login error state)
- TranslationConfig (translation display config)

Cross-module smart cast fixes for:
- LoginScreen.kt, SignUpScreen.kt (ParamsErrorMsg.params)
- TranslatableRichTextViewer.kt (sourceLang/targetLang)

Part of the KMP iOS migration tracked in #2238.
