+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedOff/login/LoginErrorManager.kt
@@ -20,36 +20,5 @@
  */
 package com.vitorpamplona.amethyst.ui.screen.loggedOff.login
 
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.setValue
-
-class LoginErrorManager {
-    interface IErrorMsg
-
-    class SingleErrorMsg(
-        val errorResId: Int,
-    ) : IErrorMsg
-
-    class ParamsErrorMsg(
-        val errorResId: Int,
-        val params: Array<out String>,
-    ) : IErrorMsg
-
-    var error by mutableStateOf<IErrorMsg?>(null)
-
-    fun clearErrors() {
-        error = null
-    }
-
-    fun error(resourceId: Int) {
-        error = SingleErrorMsg(resourceId)
-    }
-
-    fun error(
-        resourceId: Int,
-        vararg params: String,
-    ) {
-        error = ParamsErrorMsg(resourceId, params)
-    }
-}
+// Re-export from commons for backwards compatibility
+typealias LoginErrorManager = com.vitorpamplona.amethyst.commons.ui.screens.loggedOff.login.LoginErrorManager
